[SIL] [Parser] Move SILParserTUState into SILParserState.h
Moves the declaration of SILParserTUState into a new private header file: SILParserState.h
